o	Moved formatMinutes function to gameUtils o	Turned GameRow into a component ( the statistics of recent games, aka Game by Game stats.) o	PlayerProfile also turned into a component (biometrics and averages). o	Styles moved to separate file.
