+++ app/Styles/StandingsStyles.ts
